Optimizer and Code Generator now working
We have a fully working compiler! However, the optimizer could definitely be improved. It doesn't really reduce any of the redundant logic, it just cleans up the code and replaces variables with the correct registers that represent their values. Also downloaded the sample compiler (example_compiler) that Artzi had in the Padlet just to take a look at some things.