Fully Working Compiler: UPDATE
All requirements are now met including all of the error checking stuff, and writing to the files for the TAC, optimized TAC, and final MIPS. The only issue is that the error checking with unrecognized symbols or missing semicolons don't tell the user the correct line for some reason. That's the last thing to fix before this thing is completely ready!