Functions without parameters fully workinggit add -u .!
Supports functions without parameters completely, including order that they are called, scoping, and all functionality that we had before. It correctly checks for the scope of a variable when being referenced, and gets the correct value. The symbol table is updated according to scope, and supports all existing functionality. The functions can support anything we had in previous versions of the compiler (arrays, floats, order of operations, etc.), and the TAC gets put in order that the functions are called in. We also support global variables and expressions that are not a part of a function, but they must come after function definitions.